Don't pass 'none' to WebKitCSSMatrix constructor
This fixes an issue in some versions of Webkit. On affected versions a
DOM error 12 (syntax error) is thrown when 'none' is passed as the
argument to WebKitCSSMatrix's constructor. A commit fixing this issue
in Chromium is at
https://chromium.googlesource.com/chromium/blink/+/6d50d9cd8b1df3190d67fbc4767a6c466d8dcbf9^!/

In particular, I see this issue in the stock browser on a Samsung
GT-P6210 tablet.